pyproject.toml
[project]
name = "pico"
version = "0.1.0"
description = "Raspberry Pi Pico soil moisture sensor with serial UI visualization."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mpremote>=1.25.0",
    "pyserial>=3.5",
    "rich>=14.0.0",
]
authors = [
    { name = "Ishan Kaim" }
]
